Make block multi-delete conditional
For garbage collection mechanism it's beneficial to be able to
check some condition against LMDB storage in the same DB transaction
that is responsible for removal of the block.

Function `DeleteManyIf` was added which takes a predicate that is
parametrized by LMDB transaction and key that is considered for
deletion.